defense task

+++ Sharky/MicroTasks/IMicroTask.cs
@@ -6,7 +6,7 @@ namespace Sharky.MicroTasks
 {
     public interface IMicroTask
     {
-        int Priority { get; }
+        float Priority { get; }
         List<UnitCommander> UnitCommanders { get; set; }
         void ClaimUnits(ConcurrentDictionary<ulong, UnitCommander> commanders);
         IEnumerable<Action> PerformActions(int frame);
